rename the member details components
