+++ lib/widgets/pages/member_details/member_details_page.dart
@@ -12,8 +12,8 @@ import 'package:weforza/widgets/custom/dialogs/delete_item_dialog.dart';
 import 'package:weforza/widgets/custom/profile_image/async_profile_image.dart';
 import 'package:weforza/widgets/pages/add_device/add_device_page.dart';
 import 'package:weforza/widgets/pages/edit_member_page.dart';
-import 'package:weforza/widgets/pages/memberDetails/memberActiveToggle.dart';
-import 'package:weforza/widgets/pages/memberDetails/memberDevicesList/memberDevicesList.dart';
+import 'package:weforza/widgets/pages/member_details/member_active_toggle.dart';
+import 'package:weforza/widgets/pages/member_details/member_devices_list/member_devices_list.dart';
 import 'package:weforza/widgets/platform/cupertinoIconButton.dart';
 import 'package:weforza/widgets/platform/platformAwareWidget.dart';
 import 'package:weforza/widgets/providers/reloadDataProvider.dart';
