+++ lib/widgets/pages/member_details/member_devices_list/member_devices_list_item.dart
